switch back to C rand guess I don't know how to properly seed a mt19937

+++ tanks-game/rng.h
@@ -7,7 +7,7 @@ private:
 	static std::mt19937 generator;
 	static std::uniform_int_distribution<int> distribution;
 public:
-	static void Initialize(unsigned int seed);
+	static void Initialize(long long seed);
 	static double randFunc(); //[0,1)
 	static double randFunc2(); //[0,1]
 };
